Add Splash Screen with Animation
Implemented a splash screen with animations using Flutter. The splash screen includes a background image and a logo that scales, fades, and slides into place. The `GetMaterialApp` from the `get` package is used for navigation, and the splash screen is set as the home screen in the `main.dart` file. Added the `get` package to `pubspec.yaml` and `pubspec.lock`.